Consolidate websocket server into the same server
