+++ package.json
@@ -24,6 +24,7 @@
     "express": "^4.17.1",
     "express-session": "^1.17.0",
     "express-sessions": "^1.0.6",
+    "express-ws": "^4.0.0",
     "fs": "0.0.1-security",
     "multer": "^1.4.2",
     "pg": "^7.18.2",
